add likesview and split playlistsscreen
